+++ packages/eslint-plugin/rules/index.ts
@@ -1,0 +1,7 @@
+import noProcessEnvComparisonRule from './no-process-env-comparison'
+
+const rules = {
+  'no-process-env-comparison': noProcessEnvComparisonRule,
+}
+
+export default rules