+++ packages/graphql/config/environment.ts
@@ -1,6 +1,5 @@
 import { useMemo } from 'react'
 
-import { templateEnv } from '@baseapp-frontend/utils/functions/env'
 import { baseAppFetch } from '@baseapp-frontend/utils/functions/fetch/baseAppFetch'
 import { getToken } from '@baseapp-frontend/utils/functions/token/getToken'
 
@@ -68,7 +67,7 @@ export async function httpFetch(
 ): Promise<GraphQLResponse> {
   const fetchOptions = getFetchOptions({ request, variables, uploadables })
   const response = await baseAppFetch('', {
-    baseUrl: templateEnv.NEXT_PUBLIC_RELAY_ENDPOINT,
+    baseUrl: process.env.NEXT_PUBLIC_RELAY_ENDPOINT,
     decamelizeRequestBodyKeys: false,
     decamelizeRequestParamsKeys: false,
     camelizeResponseDataKeys: false,
@@ -92,7 +91,7 @@ export async function httpFetch(
 }
 
 const wsClient = createClient({
-  url: templateEnv.NEXT_PUBLIC_WS_RELAY_ENDPOINT as string,
+  url: process.env.NEXT_PUBLIC_WS_RELAY_ENDPOINT as string,
   connectionParams: () => {
     const Authorization = getToken()
     if (!Authorization) return {}
@@ -176,8 +175,8 @@ function initEnvironment() {
 }
 
 export function useEnvironment() {
-  const environment = useMemo(() => initEnvironment(), [relayEnvironment])
-  return environment
+  const env = useMemo(() => initEnvironment(), [relayEnvironment])
+  return env
 }
 
 export function getCacheByEnvironment(environment: Environment) {