+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @baseapp-frontend/config
 
+## 2.1.7
+
+### Patch Changes
+
+- Add `@baseapp-frontend/eslint-plugin` plugin.
+
 ## 2.1.6
 
 ### Patch Changes